django_task/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.7
djangorestframework==3.12.4
idna==3.2
pytz==2021.1
requests==2.26.0
simplejson==3.17.5
sqlparse==0.4.1
urllib3==1.26.6
names==0.3.0